Fixes #1, $checked_subscriptions must be array
Fix for https://github.com/Prospress/woocommerce-subscriptions-recalculate-totals/issues/1

Ran into the same problem. He misdiagnosed the problem, but it does exist. After a reset the option is filled with an empty string, which means the default fallback array is not used.

Then on line 117 of the original (121 after this commit), this string is used as an array. To fix this we must make sure the `$checked_subscriptions` variable is an array that we can manipulate.