Update tax strings
Without removing `tax` order items the new tax was correctly calculated, but the tax name remained the same. This fixes it. Background: In Germany we have a tax change. For instance, the reduced tax named "VAT 7%"  with 7% was reduced to 5%. Without this fix the tax would be correctly re-calculated but the VAT string would still say "VAT 7%" though it was already changed to "VAT 5%" in the settings.